requirementes.txt
pytest
requests
pyyaml
Flask
